Move DNS-related and Gardener-specific utilties from common to pkg/utils/gardener
