+++ pkg/operation/common/types.go
@@ -29,22 +29,6 @@ const (
 	// allow deleting the resource (if the annotation is not set any DELETE request will be denied).
 	ConfirmationDeletion = "confirmation.gardener.cloud/deletion"
 
-	// DNSProvider is the key for an annotation on a Kubernetes Secret object whose value must point to a valid
-	// DNS provider.
-	DNSProvider = "dns.gardener.cloud/provider"
-
-	// DNSDomain is the key for an annotation on a Kubernetes Secret object whose value must point to a valid
-	// domain name.
-	DNSDomain = "dns.gardener.cloud/domain"
-
-	// DNSIncludeZones is the key for an annotation on a Kubernetes Secret object whose value must point to a list
-	// of zones that shall be included.
-	DNSIncludeZones = "dns.gardener.cloud/include-zones"
-
-	// DNSExcludeZones is the key for an annotation on a Kubernetes Secret object whose value must point to a list
-	// of zones that shall be excluded.
-	DNSExcludeZones = "dns.gardener.cloud/exclude-zones"
-
 	// EtcdEncryptionSecretName is the name of the shoot-specific secret which contains
 	// that shoot's EncryptionConfiguration. The EncryptionConfiguration contains a key
 	// which the shoot's apiserver uses for encrypting selected etcd content.
@@ -103,21 +87,6 @@ const (
 
 	// AlertManagerPrefix is a constant for a prefix used for the AlertManager instance.
 	AlertManagerPrefix = "au"
-
-	// IngressPrefix is the part of a FQDN which will be used to construct the domain name for an ingress controller of
-	// a Shoot cluster. For example, when a Shoot specifies domain 'cluster.example.com', the ingress domain would be
-	// '*.<IngressPrefix>.cluster.example.com'.
-	IngressPrefix = "ingress"
-
-	// APIServerPrefix is the part of a FQDN which will be used to construct the domain name for the kube-apiserver of
-	// a Shoot cluster. For example, when a Shoot specifies domain 'cluster.example.com', the apiserver domain would be
-	// 'api.cluster.example.com'.
-	APIServerPrefix = "api"
-
-	// InternalDomainKey is a key which must be present in an internal domain constructed for a Shoot cluster. If the
-	// configured internal domain already contains it, it won't be added twice. If it does not contain it, it will be
-	// appended.
-	InternalDomainKey = "internal"
 
 	// CoreDNSDeploymentName is the name of the coredns deployment.
 	CoreDNSDeploymentName = "coredns"
