+++ pkg/operation/common/utils.go
@@ -441,37 +441,6 @@ func DeleteGrafanaByRole(ctx context.Context, k8sClient kubernetes.Interface, na
 	return nil
 }
 
-// GetDomainInfoFromAnnotations returns the provider and the domain that is specified in the give annotations.
-func GetDomainInfoFromAnnotations(annotations map[string]string) (provider string, domain string, includeZones, excludeZones []string, err error) {
-	if annotations == nil {
-		return "", "", nil, nil, fmt.Errorf("domain secret has no annotations")
-	}
-
-	if providerAnnotation, ok := annotations[DNSProvider]; ok {
-		provider = providerAnnotation
-	}
-
-	if domainAnnotation, ok := annotations[DNSDomain]; ok {
-		domain = domainAnnotation
-	}
-
-	if includeZonesAnnotation, ok := annotations[DNSIncludeZones]; ok {
-		includeZones = strings.Split(includeZonesAnnotation, ",")
-	}
-	if excludeZonesAnnotation, ok := annotations[DNSExcludeZones]; ok {
-		excludeZones = strings.Split(excludeZonesAnnotation, ",")
-	}
-
-	if len(domain) == 0 {
-		return "", "", nil, nil, fmt.Errorf("missing dns domain annotation on domain secret")
-	}
-	if len(provider) == 0 {
-		return "", "", nil, nil, fmt.Errorf("missing dns provider annotation on domain secret")
-	}
-
-	return
-}
-
 // ReadServiceAccountSigningKeySecret reads the signing key secret to extract the signing key.
 // It errors if there is no value at ServiceAccountSigningKeySecretDataKey.
 func ReadServiceAccountSigningKeySecret(secret *corev1.Secret) (string, error) {
