+++ plugin/pkg/shoot/dns/admission_test.go
@@ -22,7 +22,7 @@ import (
 	"github.com/gardener/gardener/pkg/apis/core"
 	v1beta1constants "github.com/gardener/gardener/pkg/apis/core/v1beta1/constants"
 	coreinformers "github.com/gardener/gardener/pkg/client/core/informers/internalversion"
-	"github.com/gardener/gardener/pkg/operation/common"
+	gutil "github.com/gardener/gardener/pkg/utils/gardener"
 	. "github.com/gardener/gardener/plugin/pkg/shoot/dns"
 
 	. "github.com/onsi/ginkgo"
@@ -73,8 +73,8 @@ var _ = Describe("dns", func() {
 						v1beta1constants.GardenRole: v1beta1constants.GardenRoleDefaultDomain,
 					},
 					Annotations: map[string]string{
-						common.DNSDomain:   domain,
-						common.DNSProvider: defaultDomainProvider,
+						gutil.DNSDomain:   domain,
+						gutil.DNSProvider: defaultDomainProvider,
 					},
 				},
 			}
